requirements.txt
aster-connector-python==1.1.0
flask==2.0.1
werkzeug==2.0.3
pyyaml==6.0
psutil==5.9.0